Post.tsxのコンポーネントファイルを作成しApp.tsxから引越し
